+++ packages/addon/lua/gelly/api/gx-settings-sync.lua
@@ -1,3 +1,5 @@
+local logging = include("gelly/logging.lua")
+
 -- Synchronizes the binary module with the settings
 gellyx = gellyx or {}
 assert(gellyx.settings, "Cannot synchronize settings without settings module loaded")
@@ -6,9 +8,10 @@ local BINARY_MODULE_RELATED_SETTINGS = {
 	"smoothness",
 	"simulation_substeps",
 	"simulation_iterations",
+	"max_particles"
 }
 
-function gellyx.settings.updateBinaryModuleSettings()
+function gellyx.settings.updateBinaryModuleSettings(changedConvar)
 	-- The binary module's settings are ephemeral, which is why we synchronize them with the convars
 	gelly.SetGellySettings({
 		FilterIterations = gellyx.settings.get("smoothness"):GetInt(),
@@ -19,6 +22,15 @@ function gellyx.settings.updateBinaryModuleSettings()
 		Substeps = gellyx.settings.get("simulation_substeps"):GetInt(),
 		Iterations = gellyx.settings.get("simulation_iterations"):GetInt(),
 	})
+
+	if changedConvar == gellyx.settings.getFullName("max_particles") then
+		-- This is one of our most expensive changes so we ensure it's only done when necessary
+		gelly.ChangeMaxParticles(gellyx.settings.get("max_particles"):GetInt())
+		gellyx.presets.select(gellyx.presets.getActivePreset().Name)
+		hook.Run("GellyRestarted")
+
+		logging.warn("Max particles set to " .. gellyx.settings.get("max_particles"):GetInt() .. "!")
+	end
 end
 
 gellyx.settings.registerMultipleOnChange(BINARY_MODULE_RELATED_SETTINGS, gellyx.settings.updateBinaryModuleSettings)